% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_profiles.R
\name{cluster_profiles}
\alias{cluster_profiles}
\title{K-means Clustering of OUtcome on Exposure profiles}
\usage{
cluster_profiles(wdata, center = TRUE, scale = TRUE, k_count = 20)
}
\arguments{
\item{wdata}{a data frame of predicted exposure values, such as those derived from the function gam_predict()}
}
\value{
a list of kmeans values, and plots
}
\description{
This function takes a matrix of outcome on exposure predicted values and performs a kmeans clustering to aid in identifying similar profiles
}
\examples{
cluster_profiles()
}
\keyword{GAM}
