% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_predict.R
\name{gam_predict}
\alias{gam_predict}
\title{Generalized additive model fitting}
\usage{
gam_predict(
  wdata,
  outcome = NA,
  exposure = NA,
  linear_covariates = NA,
  smooth_covariates = NA,
  rnt_outcome = FALSE,
  return_prediction_only = TRUE,
  prediction_n = 25,
  outlier_cutoff = 5,
  outlier_method = "iqr",
  new_pred_data_frame = NA,
  messages = FALSE
)
}
\arguments{
\item{wdata}{a data frame passed to function containing necessary data for analysis.}

\item{outcome}{a single string character of the column name for the dependent or response variable.}

\item{exposure}{a single string character of the column name for the independent or explanatory variable.}

\item{linear_covariates}{a vector of string(s) that are also column names used to define variables that will be set as parametric (linear) covariates.}

\item{smooth_covariates}{a vector of string(s) that are also column names used to define variables that will be set as non-linear (smooth, s()) covariates.}

\item{return_prediction_only}{binary TRUE or FALSE, defaulted to TRUE, asking if only a vector of exposure predicted values should be returned or a data frame of predicted values and the variable data used in the prediction.}

\item{prediction_n}{how many values should be predicted across the range of the exposure? Default is 25.}

\item{new_pred_data_frame}{a data matrix containing the model independent and covariates from which exposure values should be predicted. If left as NA, an equal distant range of dependent values of length 25, and the mean of each numeric and max of each factor will be used in the prediction matrix.}

\item{messages}{should verbose messages be printed to screen}

\item{rnt_dependent}{binary TRUE or FALSE if the dependent or response variable should be rank normal transformed.}
}
\value{
returns a list object containing two objects, (1) 'fit' which holds the gam() model and (2) 'summary' which is a vector of summary statistics derived from the model.
}
\description{
This function fits a GAM to the data
}
\examples{
gam_predict()
}
\keyword{GAM}
