% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glsmr.R
\name{glsmr}
\alias{glsmr}
\title{GAM and linear-strata Mendelian randomization}
\usage{
glsmr(
  wdata,
  outcome = NA,
  exposure = NA,
  instrument = NA,
  linear_covariates = NA,
  smooth_covariates = NA,
  strata = 4,
  rnt_outcome = FALSE,
  weights_variable = NA,
  sd_outlier_cutoff = 5,
  sd_or_iqr_cutoff = "iqr",
  messages = FALSE
)
}
\arguments{
\item{wdata}{a data frame passed to function containing necessary data for analysis}

\item{outcome}{a single string character of the column name for the outcome or dependent or response variable}

\item{exposure}{a single string character of the column name for the exposure or independent or explanatory variable}

\item{instrument}{a data frame passed to function containing necessary data for analysis}

\item{linear_covariates}{a vector of string(s) that are also column names used to define variables that will be set as parametric (linear) covariates.}

\item{smooth_covariates}{a vector of string(s) that are also column names used to define variables that will be set as non-linear (smooth, s()) covariates.}

\item{strata}{a string or vector to define how strata should be defined.
* `strata = quartiles`: automatically identifies quartiles or 4 equally sized strata
* `strata = decile`: automatically identifies deciles or 10 equally sized strata
* `strata = c(1,10,20,30)`: a user defined numeric vector to define boundaries for each strata.
   - The numeric vector example provided will define 4 strata. Lower bound values are inclusive, upper bounds are exclusive, to the exception of the last bounding value.}

\item{rnt_outcome}{binary TRUE or FALSE if the dependent or response variable should be rank normal transformed.}

\item{weights_variable}{a single string character of the column name for a weights variable}

\item{sd_outlier_cutoff}{a single numeric value to define a cutoff value for how many iqr or sd units outlier values}

\item{sd_or_iqr_cutoff}{a single string character of "iqr" or "sd" to determine if outlier should be determined by means and sd or medians and iqr.}

\item{messages}{should a progress message be printed to screen - binary TRUE or FALSE}
}
\value{
returns a glsmr object containing the complete linear and GAM models for the full data set, summary statistics for the data, a strata observational table, and a strata TSLS (MR) table.
}
\description{
This function derives strata defined observational and TSLS (MR) estimates using linear modeling.
The full data set will be fit to a linear model and a generalized linear model (GAM).
Then subsequently the exposure data will be stratified and a linear model will be used to derive
effect estimates for each strata.
}
\examples{
glsmr()
}
\keyword{GAM}
\keyword{MR}
\keyword{strata}
\keyword{stratified}
