% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimates.R
\name{iv_estimates}
\alias{iv_estimates}
\title{linear model fitting to estimate the instrument on dependent effect coefficient (beta_ie) and the proportion of variance explained (eta-squared) in dependent by instrument}
\usage{
iv_estimates(wdata, dependent, instrument, covariates)
}
\arguments{
\item{wdata}{a data frame passed to function containing necessary data for analysis}

\item{dependent}{a single string character of the column name for the dependent or response (y) variable.}

\item{instrument}{a single string character of the column name for the instrument, which will be set as the independent or explanatory variable in the model.}

\item{covariates}{a character vector that are also column names used to define variables that will be set as covariates.}
}
\value{
returns a lmfit object that is a list containing (1) 'fit' a lm() object and (2) 'summary' a vector of summary statistics.
}
\description{
This function estimates the instrument on dependent effect coefficient (beta) and the proportion of variance explained (eta-squared) in dependent by instrument
}
\examples{
iv_estimates()
}
\keyword{linear}
\keyword{model}
