% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_free_exposure.R
\name{iv_free_exposure}
\alias{iv_free_exposure}
\title{linear model fitting to estimate exposure values in the absence of the instrument effect}
\usage{
iv_free_exposure(
  wdata,
  exposure,
  instrument,
  covariates = NULL,
  exposure_mean_normalize = TRUE
)
}
\arguments{
\item{wdata}{a data frame passed to function containing necessary data for analysis}

\item{exposure}{a single string character of the column name for the exposure variable, which will be set as the response variable in this model.}

\item{instrument}{a single string character of the column name for the instrument, which will be set as the independent or explanatory variable in the model.}

\item{covariates}{a character vector that are also column names used to define variables that will be set as covariates.}

\item{exposure_mean_normalize}{binary TRUE or FALSE. TRUE statements will re-center the residulas to have a mean value of that observed for the exposure.}
}
\value{
returns the data frame provided but with an additional column 'iv_free_exposure'.
}
\description{
This function estimates exposure values in the absence of the instrument effect by fitting a linear model of instrument + covariates on exposure and extracting the residuals.
}
\examples{
iv_free_exposure()
}
\keyword{linear}
\keyword{model}
