% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_predicted_exposure.R
\name{iv_predicted_exposure}
\alias{iv_predicted_exposure}
\title{linear model fitting to estimate the instrument predicted exposure, sometimes referred to as d.hat in an MR framework.}
\usage{
iv_predicted_exposure(wdata, exposure, instrument, covariates = NULL)
}
\arguments{
\item{wdata}{a data frame passed to function containing necessary data for analysis}

\item{exposure}{a single string character of the column name for the exposure variable, which will be set as the response variable in this model.}

\item{instrument}{a single string character of the column name for the instrument, which will be set as the independent or explanatory variable in the model.}

\item{covariates}{a character vector that are also column names used to define variables that will be set as covariates.}
}
\value{
returns a lmfit object that is a list containing (1) 'fit' a lm() object and (2) 'summary' a vector of summary statistics.
}
\description{
This function estimates instrument predicted exposure by fitting a linear model of instrument + covariates on exposure and extracting the fitted values.
}
\examples{
iv_predicted_exposure()
}
\keyword{linear}
\keyword{model}
