% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivregfit.R
\name{ivregfit}
\alias{ivregfit}
\title{MR (ivreg) model fitting}
\usage{
ivregfit(
  wdata,
  outcome,
  exposure,
  instrument,
  covariates = NA,
  weights_variable = NA,
  rnt_outcome = FALSE
)
}
\arguments{
\item{wdata}{a data frame passed to function containing necessary data for analysis}

\item{outcome}{a single string character of the column name for the outcome or dependent or response variable}

\item{exposure}{a single string character of the column name for the exposure or independent or explanatory variable}

\item{instrument}{a data frame passed to function containing necessary data for analysis}

\item{covariates}{a character vector that are also column names used to define variables that will be set as covariates.}

\item{weights_variable}{a single string character of the column name for a weights variable}

\item{rnt_outcome}{binary TRUE or FALSE if the dependent or response variable should be rank normal transformed.}
}
\value{
returns a ivreg object that is a list containing (1) 'fit' a ivreg() object and (2) 'summary' a vector of summary statistics.
}
\description{
This function fits instrumental variable (MR, ivreg) models
}
\examples{
ivregfit()
}
\keyword{MR}
\keyword{model}
