% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmfit.R
\name{lmfit}
\alias{lmfit}
\title{linear model fitting}
\usage{
lmfit(wdata, dependent, independent, covariates, rnt_dependent = FALSE)
}
\arguments{
\item{wdata}{a data frame passed to function containing necessary data for analysis}

\item{dependent}{a single string character of the column name for the dependent or response variable}

\item{independent}{a single string character of the column name for the independent or explanatory variable}

\item{covariates}{a character vector that are also column names used to define variables that will be set as covariates.}

\item{rnt_dependent}{binary TRUE or FALSE if the dependent or response variable should be rank normal transformed.}
}
\value{
returns a lmfit object that is a list containing (1) 'fit' a lm() object and (2) 'summary' a vector of summary statistics.
}
\description{
This function fits observational linear models
}
\examples{
lmfit()
}
\keyword{linear}
\keyword{model}
