% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normW.R
\name{normW}
\alias{normW}
\title{Estimate shapiro-wilk's W statistic for normality}
\usage{
normW(y)
}
\arguments{
\item{y}{a numeric vector from which Shapiro-Wilk test of normality will be performed}
}
\value{
an estimate of normality, the W-statistic.
}
\description{
This function estimates the Shapiro-Wilk's W statistic for normality
}
\examples{
normW()
}
\keyword{Shapiro-Wilk}
\keyword{W-statistic}
\keyword{normality}
