% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rntransform.R
\name{rntransform}
\alias{rntransform}
\title{rank normal tranformation of a vector}
\usage{
rntransform(y, split_ties = TRUE)
}
\arguments{
\item{y}{a numeric vector which will be rank normal transformed}

\item{split_ties}{a binary string of TRUE (default) or FALSE indicating if tied values, of the same rank, should be randomly split giving them unique ranks.}
}
\value{
returns vector of rank normalized data
}
\description{
This function rank normal transforms a vector of data. The procedure is built off of that provided in the GenABEL pacakge.
}
\examples{
rntransform()
}
\keyword{normal}
\keyword{rank}
\keyword{transformation}
