% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_data.R
\name{stratify_data}
\alias{stratify_data}
\title{stratify data}
\usage{
stratify_data(wdata, stratify_on, strata = 4)
}
\arguments{
\item{wdata}{a data frame passed to function containing necessary data for analysis}

\item{stratify_on}{column name to identify the variable to stratify on}

\item{strata}{a string or vector to define how strata should be defined.Defaulted to 4 bins or quartiles.}
}
\value{
returns
}
\description{
This function stratifies the working data set into either quantiles or user defined bins
}
\examples{
stratify_data()
}
\keyword{stratify}
