% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_ivratio.R
\name{stratify_ivratio}
\alias{stratify_ivratio}
\title{estimates the causal effect estimate or beta_iv for strata}
\usage{
stratify_ivratio(strata_beta_ie, strata_beta_io, beta_ie, tsls_ss)
}
\arguments{
\item{strata_beta_ie}{stratified instrument on exposure estimates from stratify_lmfit()}

\item{strata_beta_io}{stratified instrument on outcome estimates from stratify_lmfit()}

\item{beta_ie}{the instrument on exposure effect estimate for the entire data set}

\item{tsls_ss}{the two-stage least square summary table from the full data set}
}
\value{
returns
}
\description{
This function estimates the causal effect estimate or beta_iv for strata
}
\examples{
stratify_ivratio()
}
\keyword{stratify}
