% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_lmfit.R
\name{stratify_lmfit}
\alias{stratify_lmfit}
\title{runs lmfit() on output of stratify_data()}
\usage{
stratify_lmfit(wdata, outcome, exposure, covariates)
}
\arguments{
\item{wdata}{a strata_data() object}

\item{outcome}{a single string character of the column name for the outcome or response variable}

\item{exposure}{a single string character of the column name for the exposure or explanatory variable}

\item{covariates}{a character vector that are also column names used to define variables that will be set as covariates.}
}
\value{
returns
}
\description{
This function stratifies the working data set into either quantiles or user defined bins
}
\examples{
stratify_lmfit()
}
\keyword{stratify}
