% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_sumstats.R
\name{stratify_sumstats}
\alias{stratify_sumstats}
\title{summary statistics for the exposure within each strata}
\usage{
stratify_sumstats(wdata, exposure, model_vars = c(outcome, exposure))
}
\arguments{
\item{wdata}{a data frame passed to function containing necessary data for analysis}

\item{exposure}{column name to summarize across strata}
}
\value{
returns a data frame of summary statistics on the column of interest for each strata
}
\description{
This function estimates a few useful summary statistics for the exposure within each strata
}
\examples{
stratify_sumstats()
}
\keyword{statitics}
\keyword{stratify}
\keyword{summary}
