% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamfit.R
\name{gamfit}
\alias{gamfit}
\title{Generalized additive model fitting}
\usage{
gamfit(
  wdata,
  dependent = NA,
  independent = NA,
  linear_covariates = NA,
  smooth_covariates = NA,
  rnt_dependent = FALSE
)
}
\arguments{
\item{wdata}{a data frame passed to function containing necessary data for analysis.}

\item{dependent}{a single string character of the column name for the dependent or response variable.}

\item{independent}{a single string character of the column name for the independent or explanatory variable.}

\item{linear_covariates}{a vector of string(s) that are also column names used to define variables that will be set as parametric (linear) covariates.}

\item{smooth_covariates}{a vector of string(s) that are also column names used to define variables that will be set as non-linear (smooth, s()) covariates.}

\item{rnt_dependent}{binary TRUE or FALSE if the dependent or response variable should be rank normal transformed.}
}
\value{
returns a list object containing two objects, (1) 'fit' which holds the gam() model and (2) 'summary' which is a vector of summary statistics derived from the model.
}
\description{
This function fits a GAM to the data
}
\examples{
gamfit()
}
\keyword{GAM}
