% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_outliers.R
\name{id_outliers}
\alias{id_outliers}
\title{Identify outliers using mean and SD or median and interquartile range}
\usage{
id_outliers(y, outlier_method = "iqr", outlier_cutoff = 5)
}
\arguments{
\item{y}{a numeric vector from which outlier indexes will be identified.}

\item{outlier_method}{a single string character of "iqr" or "sd" to determine if outlier should be determined by means and sd or medians and iqr.}

\item{outlier_cutoff}{a single numeric value to define a cutoff value for how many iqr or sd units outlier values.}
}
\value{
returns vector indicating the index of an outlier(s).
}
\description{
This function identifies outliers using mean and SD or median and interquartile range
}
\examples{
id_outliers()
}
\keyword{IQR}
\keyword{SD}
\keyword{mean}
\keyword{median}
\keyword{outliers}
