% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glsmr.R
\name{plot_glsmr}
\alias{plot_glsmr}
\title{Plot functions for glsmr objects}
\usage{
plot_glsmr(glsmr_obj)
}
\arguments{
\item{glsmr_obj}{a glsmr object from the function glsmr()}
}
\value{
ggplot figure
}
\description{
This function makes a few useful plots for glsmr objects
}
\examples{
plot_glsmr()
}
\keyword{glsmr}
\keyword{plot}
